fix layout ui
