+++ src/app/app.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ConfirmationService, MessageService } from 'primeng/api';
 import { Product } from './domain/product';
 import { ProductService } from './services/productservice';
+import { SidebarService } from './main/layout/api/services/sidebar.service';
 
 @Component({
     selector: 'app-root',
@@ -10,10 +11,15 @@ import { ProductService } from './services/productservice';
     providers: [ConfirmationService, MessageService, ProductService]
 })
 export class AppComponent implements OnInit {
-
-    constructor() { }
+    darkMode: boolean;
+    constructor(private sidebarService: SidebarService) { }
 
     ngOnInit(): void {
+        this.sidebarService.getMode().subscribe({
+            next: (response) => {
+                this.darkMode = response;
+            }
+        });
 
     }
 }
