+++ src/app/main/layout/app-layout.component.html
@@ -1,6 +1,6 @@
-<div class="layout-wrapper " [ngClass]="darkMode? 'bodyDark': 'bodyLight'">
+<div class="layout-wrapper">
   <div class="layout-main-container">
-    <div class="layout-main" [ngStyle]="{}" [ngClass]="!isCollapsed ? 'body-trimmed' : 'body-md-screen' ">
+    <div class="layout-main" [ngClass]="!isCollapsed ? 'body-trimmed' : 'body-md-screen' ">
       <router-outlet></router-outlet>
     </div>
   </div>
