+++ src/app/main/layout/app-layout.component.scss
@@ -1,11 +1,3 @@
-.bodyDark {
-    background: #22577a;
-}
-
-.bodyLight {
-    background: #f8f7ff
-}
-
 .body {
     width: calc(100% - 5rem);
     height: 100vh;
@@ -16,16 +8,17 @@
     transition: all .5s ease;
     background: transparent;
     padding: 3rem;
-    background: var(--background);
 }
 
 .body-trimmed {
-    width: calc(100% - 16.5625rem);
+    height: 100%;
+    // width: calc(100% - 16.5625rem);
     margin-left: 5rem;
 }
 
 .body-md-screen {
-    width: calc(100% - 5rem);
+    // width: calc(100% - 5rem);
+    height: 100%;
     margin-left: 16.5625rem;
 
 }
