+++ src/app/main/pages/manage-items/manage-items.component.html
@@ -64,7 +64,7 @@
         </p-table>
     </div>
     
-    <p-dialog [(visible)]="productDialog" [style]="{width: '450px'}" header="Product Details" [modal]="true"
+    <p-dialog [(visible)]="productDialog" [style]="{width: '50vw'}" header="Product Details" maximizable="true" [modal]="true"
         styleClass="p-fluid">
         <ng-template pTemplate="content">
             <div class="p-field m-2">
