+++ src/app/main/pages/manage-sales/manage-sales.component.html
@@ -1,5 +1,5 @@
 <div class="px-8">
-    <div class="px-8">
+    <div class="">
         <p-dataView #dv [value]="products" [rows]="5" [paginator]="true" class="px-8">
             <ng-template let-product pTemplate="listItem">
                 <div class="col-12">
