+++ src/index.html
@@ -5,7 +5,8 @@
   <meta http-equiv="X-UA-Compatible" content="IE=edge" />
   <title>PrimeNG Quickstart CLI</title>
   <base href="/">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
+  <link href='https://fonts.googleapis.com/css?family=Sofia' rel='stylesheet' />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0, minimum-scale=1.0" />
   <link rel="icon" type="image/x-icon" href="favicon.ico">
   <script src="https://unpkg.com/feather-icons"></script>
 </head>
