+++ src/styles.css
@@ -1,9 +1,9 @@
 /* You can add global styles to this file, and also import other style files */
-:root {
+/* :root {
     --background-color: #006d77;
     --background: #000000;
     --sidebar-background: #ededed;
-}
+} */
 
 * {
     margin: 0;
@@ -14,8 +14,8 @@
 body {
     position: relative;
     color: #ffffff;
-    background: var(--background-color);
-    font-family: Georgia, 'Times New Roman', Times, serif;
+    /* background: var(--background-color); */
+    font-family: Arial, Helvetica, sans-serif;
 }
 
 .sidenav {
@@ -133,4 +133,16 @@ body {
     .sidenav-link-text {
         color: green;
     }
+}
+
+.bodyDark {
+    background: #22577a;
+    /* width: 100%;
+    height: 100%; */
+}
+
+.bodyLight {
+    background: #f8f7ff;
+    /* width: 100%;
+    height: 100%; */
 }
